Full canvas file and node support & polishing
# Version 1.2
## New Features
Text Node Editing: Select any text card on the canvas and edit it in the side window. Changes sync instantly in both directions.

Convert to File: Added a button (and hotkey Mod+Shift+F) to instantly convert a temporary text card into a permanent Markdown file. The plugin handles the file creation and automatically replaces the node on the canvas.

Settings Improvements:
    Direct link to Obsidian Hotkeys settings.
    Options to clean up temporary files.
    New "Support Development" section.

## Expanded File Support
The plugin is no longer limited to just Markdown files. Full range of Obsidian media and node types:
    Node: Canvas Node (Text Cards)
    Bases: .base
    JSON Canvas: .canvas
    Images: .avif, .bmp, .gif, .jpeg, .jpg, .png, .svg, .webp
    Audio: .flac, .m4a, .mp3, .ogg, .wav, .webm, .3gp
    Video: .mkv, .mov, .mp4, .ogv, .webm
    PDF: .pdf